// src/routes/darkpearl/types.ts
export type GitHubRepoLink = {
	owner: string
	name: string
	full_name: string
	default_branch: string
}